+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `id_translation.logging.EMIT_LOGGED_WARNINGS`; set to `False` to disable warnings that are emitted as logs.
 
+### Changed
+- Links in warnings and exceptions now to the current version in the docs (instead of the _'stable'_ version).
+
 ## [1.2.1] - 2026-05-09
 
 ### Fixed